feat: include trainer nickname in quote list response
